revisar towerContest
Haciendo el ejercicio manualmente si funciona el anidamiento, pero el simulador no las ordena correctamente

+++ Cup.java
@@ -254,7 +254,7 @@ public class Cup {
      * @return The calculated height
      */
     private int calculateHeight(int id) {
-        return (int) 2 * id - 1;
+        return 2 * id - 1;
     }
 
     /**
